Add main method back to any files, Add new cash payment page.
Because any WindowsForm need own Main Method if don't have this in file
it can't be run as normal.

Please beware to delete main method in WindowsForm Project, You should
delete main method if that class create for use as object only.

Thank you.

PS. We need to talk about it tomorrow.